refactor: 불필요한 context api을 컴포넌트 합성으로 해결
